Add files via upload
The powerpoint describes the method of calculating the graphs connecting the atoms as nodes that develop multidimensional topological features like cavities, holes, etc and classify the structures based on the birth and death of the topological features in the feature space. Suppose the filtration parameter is chosen to be Euclidean distance between two point in the point cloud. In a denser structure, intervals between the birth and death of a void should be small. That means the persistence interval for a specific filtration value, is smaller if compared to the less dense structures.
The ppt also gives examples of HDA-Silicon, LDA-Silicon and another arbitrary interfacial Silicon structure. Similar tetrahedral structure exists in ice. The classification methodology classifies different phases of ice too. The electrical conductivity of 2D Silicon depends on the density and coordination number of Silicon.This classification aids in classificaion of amorphous semiconducting structures.